drivers:misc:ti-st: change protocol parse logic
TI shared transport driver had to specifically know the
protocol headers for each type of data it can receive to
properly re-assemble data if its fragmented during UART
transaction or fragment if the data is an assembly of

different protocol data.

Now the individual protocol drivers provide enough header
information for shared transport driver to do this in a
generic way applicable for all protocols.

Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>