usb: dwc3: handle in and out ep irq separately
Later we'll see if this is really necessary,
but for now it'll let us see the similarities
on the code once it's implemented.

There are plenty of structures which we still
don't have. Before we can implement this, we
need e.g. a structure to represent the HW
endpoint and a list or array of endpoints
in struct dwc. See musb driver for a reference.

Signed-off-by: Felipe Balbi <balbi@ti.com>