usb: dwc3: fixup event buffer definition
so we have 32 1kib event buffers. Each of those 1kib buffers holds space
for around 256 entries. Each buffer has entries like size, count and
address. The whole thing is used in a ring buffer fashion and it fills one
entry after the other. It seems that once one 1kib is complete then it
goes to the next one.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>