usb: dwc3: reorganize event buffer handling
The list is replaced with an array because new event entries are attached
to the tail so we need quick access to the buffer.
Right know I don't know how and when the core iterates from one event buffer to
the next so we check all of them to see if there is actually any work to
do.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>