usb: dwc3: add core deinit function
It should do the opposite of dwc3_core_init() which is currently just
freeing event buffers which is now also used in error case (if the
gadget thingy goes wrong).
I'm not really sure if this is the proper way to deinit event buffers.
However since the size is 0 it should not write to main memory anymore.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>