usb: dwc3: add disconnect interrupt handling
So. The DWC3_DCTL_INITU[12]ENA block is commented out because we care
about power saving (much) later.

According to the 9.1.6 we don't have to do anything if a disconnect
event occurred. If we remain idle then the core will send us an RESET
and CONNECT_DONE event.

If we do something, we simply initiate and disconnect. This means we
stop all active EPs and set the RUN/STOP bit to stop. I'm not sure if
"active" EPs are marked as active in the ->enable hook (which I am using
now) or during ->enqueue.

However, after I write this I think that we might want do nothing
because we need somehow to trigger tne RUN/STOP bit and I don't know how
to do that :)
Ach, DWC3_DEPCMD_CMDIOC generates a interrupt once the command is done.
However I don't know why I should be interrested in this. Probably I
must not trigger the RUN/STOP bit too early.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>