usb: dwc3: gadget: create/destroy trb_pool based on ep enable/disable
We don't need to waste memory for never used endpoints,
so we can defer creation of dma_pool to usb_ep_enable()
time and destroy it when the endpoint gets disabled
by usb_ep_disable().

From that pool, we will allocate the Transfer Request
Blocks (TRBs) and make them chained to each other.

Whenever we see a usb_ep_queue(), we can allocate a
TRB and let the Hardware know about it.

Signed-off-by: Felipe Balbi <balbi@ti.com>