usb: dwc3: add an usb control request
from what I see, the driver has to provide some room for it. MUSB reads
it from a FIFO but DWC3 wants to start an transfer and to have a buffer
waiting for it.
The buffer has to be aligned to a 16bytes boundary. Since the struct is
only 8 bytes in size it is smaller than the required boundary. So I
though it might be ecological to align th dwc struct and embedd the
request struct into it. Once we at it, also add an TRB for sending the
usb struct.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>