usb: dwc3: add first setup phase for ep0
According to the graph we have to wait for XferNotReady event and then
we have to enqueue a TRB to to retrieve the Setup packet. The reference
code has its own packet which is used for that and they enqueue it and
start the transfer. They use 0 as the transfer length which I do also.
They however allocate 8 * 5 bytes for the buffer while I do just one. I
did not find any reference if one usb control packet is enought or if we
need to reserve room for more.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>