usb: dwc3: gadget: implement wakeup method
wakeup is used to issue a remote wakeup to
the host. I'm not sure if my understanding
that we don't have remote wakeup on SuperSpeed
is correct, but that's how I understood it.

In any case, we can always simply remove the
check for SuperSpeed later, should we change
our understanding.

Signed-off-by: Felipe Balbi <balbi@ti.com>