usb: dwc3: don't init ep0/1 in gadget_init()
Right now we initialized ep0/1 in gadget_init() and in connection_done
interrupt. This seems unnecessary and we leak memory for the trb_pool.
It seems to be enough to do this in connect_done interrupt so I move it
there.
Since the init function allocates memory for the trb_pool we would have
to pass GFP_ATOMIC to it so it works. This is totally bogus since we
have a static TRB for ep0/1 so we don't need this at all.
This patch also groups the TRB pool's alloc/free functions. I have the
feeling that we have to replace it sooner or later.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>