usb: dwc3: remove irqreturn_t from most functions
irqreturn_t is used to tell the core if this interrupt was issues by our
driver or not. If it was, we have to do something to stop it. We ACK the
interrupt in dwc3_process_event_buf() by writting the number of
processed events. So even if we can't handle one specific event _or_ we
received an XferComplete but have no requests in the queue we still
noticed that this was our interrupt and we acknowledged it.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>