usb: dwc3: convert the current TRB queue design into a ring buffer
I hate doing this but I don't see any other way how to get ISOC with
that update feature working.
So this should implement a ring buffer: free_slot points to the next
available slot and busy_slot points to the slot which is currently owned
by the hardware. If both of them are equal then we have either 0 or
DWC3_TRB_NUM slots available. The arithemetic requires DWC3_TRB_NUM to
be power of two.
Right now I assume that equal means empty because we are only called on
IOC (so we completed atleast one request) or if the queue has one
request queued (so the hardware was idle).

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>