usb: dwc3: fixup Kconfig madness
we redefined USB_DWC3 and made it depend on it self. This did not
produce an error but was not correct either. Now we have:

USB_DWC3	- The controller as a whole. Without it we don't have
		  neither HOST nor DEVICE
USB_GADGET_DWC3 - selects the gadget part
USB_DWC3_GADGET - actually compiles he gadget part. Part of gadged
		  madness

While fixing this, also providing static inlines for the gadget code
incase we want to use only the HOST part without gadget support.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>