usb: dwc3: move queued requests onto a seperate list
since we attempt to queue requests while the transfer is in progress we
have to distinguish requests which have allready a TRB setup from those
which do not. The req->queued field is one way to do this but that means
we have to cycle through the list to get the first non-queued requests.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>