usb: dwc3: introduce platform_data.h
that platform_data is only used between glue
layer and core files, so it doesn't need to
be exported anywhere.

We are using that to require glue layer to
give core platform_device ioremapped memory
bases. That way we can easily cope with
different integrated versions of the core
where the memory bases are different.

Signed-off-by: Felipe Balbi <balbi@ti.com>