usb: dwc3: support for 64bit addr_t
dma_addr_t is bad because it can be 32bit or 64bit and we need a fix
value here. Merging bpl and bph also "enables" support TRBs which are
not located within the first four GiB of physical memory.
Even though I've added *le* helper the access is broken: We have to turn
around the union (for length, pcm and so on) to support BE. The easiest
part is probably to use bit shifting which also needs less access .)

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>