usb: dwc3: link the ISOC TRBs
The last TRB is now pointing to the first one so we have a ring. From
what I see in the databook the LINK-TRB is never updated by the HW so we
don't have to re-set the HWO bit for instance.
I haven't found a hint that we get an event for the LINK-TRB so we
simply skip over it in the give-back function so the ring buffer isn't
broken :)

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>