usb: dwc3: fixup bitfield definitions
Only bitfields of the same type can be merged together. Lets look at
this:

|#include <stdio.h>
|#include <stdlib.h>
|
|typedef unsigned short u16;
|
|struct blah {
|        unsigned        a:1;
|        u16             b;
|        unsigned        c:15;
|} __attribute__((packed));
|
|struct bleh {
|        unsigned        a:1;
|        unsigned        b:16;
|        unsigned        c:15;
|} __attribute__((packed));
|
|int main(void)
|{
|        struct blah have;
|        struct bleh expected;
|
|        printf("=> %ld vs %ld\n", sizeof(have), sizeof(expected));
|        return 0;
|}

and executing the code above gives us:
|=> 5 vs 4

The reason is that in blah unsigned and u16 are of different types and
therefore can not be summerized. We run into this kind of problem in
several places, I noticed it in dwc3_trb.

Therefore I converted _all_ structs into proper types no matter if they
are affected or not. Alteast I'm sure it is not broken.

I convert all of them to u32 because all of them represent a hardware
register of 32 bits in size. unsigned is ugly as hell. I know it is
"unsigned int" but just wait for an expert changing it to unsigned long.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>