usb: dwc3: make the TRB structure endian safe
I tried not to lose the bitfields within dwc3_trb while making it
endian safe so I introduced dwc3_trb_hw which represents the hw format
which is little endian.
This did not end as messy as I assumed in the first place because there
are just four functions where we touch the TRB: during preparing a TRB
and once it is done for generic EP and EP0.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>