usb: dwc3: convert to only one base
Arch or Bus-specific code will pass only one
memory base down to driver, so we need to take
care of the rest using offsets. In fact, our
offsets were defined all so that they were
increments of the first resource, so there's
no problem in simply converting everything
without changes to the offsets.

Signed-off-by: Felipe Balbi <balbi@ti.com>