usb: dwc3: use platform_device_id
We can use that to pass a lot of information
to the driver which can be used in runtime
to decide how the driver will behave.

With it, we can completely remove the glue
layer and rely on the pm_runtime being implemented.

If pm_runtime isn't implemented, we simply
don't support PM with that particular architecture.

Signed-off-by: Felipe Balbi <balbi@ti.com>