usb: dwc3: use platform_get_resource_byname
This is how lower layers are passing us the address:

static struct omap_hwmod_addr_space omap54xx_usb_otg_ss_addrs[] = {
        {
                .name           = "wrapper",
                .pa_start       = 0x4a020000,
                .pa_end         = 0x4a02ffff,
                .flags          = ADDR_TYPE_RT
        },
        {
                .name           = "dwc_usb3",
                .pa_start       = 0x4a030000,
                .pa_end         = 0x4a03ffff,
        },
};

wrapper is an OMAP-specific piece of address space
which takes care of plugging the core into OMAP
SoC. Details such as Power Management as handled
by that "wrapper".

Clearly, the dwc3 core driver only wants to touch
the dwc_usb3 part of the address, which is actually
what the Core was designed to handle, so let's use
platform_get_resource_byname() and update the resource
name on the HAPS PCI glue layer so the core still
fetches the address base.

Signed-off-by: Felipe Balbi <balbi@ti.com>