usb: dwc3: core: get IRQ by name
use the same name as for the address space to grab
the correct IRQ number. At least on OMAP, due to the
way the core is integrated into the OMAP PM Subsystem,
we have two separate IRQs and two "separate" address
spaces.

The dwc_usb3 IRQ and Address, are peculiar to the DWC
USB3 core while the "wrapper" address and IRQ are
peculiar to OMAP-only. Those will be handled on an
OMAP glue layer to be (re)introduced on a later
patch. That particular glue layer will handle stuff
such as the OMAP-specific power management details,
turn on/off of internal DMA, via DMAEnable bit on
SysConfig register, etc.

Signed-off-by: Felipe Balbi <balbi@ti.com>