usb: dwc3: gadget: don't drop run/stop on disconnect
Databook asks to drop that bit, but if we do,
we would need to go through POR sequence again.

We are doing the same thing which reference
code is doing and not dropping the run/stop
bit. Later we might decide to phase POR out
to a non-init function so that we can re-use
and complete the disconnect by dropping run/stop
again.

Signed-off-by: Felipe Balbi <balbi@ti.com>