usb: dwc3: ep0: don't queue always to ep0
we need to care about direction. This is not
the best way to handle it, but it's working
like a charm. A full enumeration has been
achieved with this patch.

It seems like we need to start thinking, again,
about combining HW endpoints into a duplex USB
(logical) endpoint representation like other
controller drivers (like MUSB) are doing. For
now, this patch starts a discussion while getting
us in a working state.

Signed-off-by: Felipe Balbi <balbi@ti.com>