usb: dwc3: don't queue req for ep0/1 if something is going on
We have one TRB and our state depends on correct handling. So we should
not enqueue and start one packet on ep1 while there is something going
on ep0. In the begin of the setup phase we are in EP0_IDLE and drop all
requests but later while processing one requests we may accidently start
processing another request.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Felipe Balbi <balbi@ti.com>