pia35x: fix powerup for expansion board
When the TiWi-R2 detects a low signal on any of it's BT/WLAN_EN inputs
for more than 10 ms, it sets the module into low power mode. The device
will only exit this mode when a new power-up sequence is initiated.

This means that a warm reset would permanetly shutdown the module, so we
have to shutdown the expansion power during startup to allow proper
reinitialization.