Move EDMA suspend-resume callbacks to late-suspend/early-resume
A race issue was seen during intense MMC testing where suspend/resume
was done along with a MMC transfer operation.

During the normal system resume process,
	- edma driver resume callback will be called first which sets
	  up the requested edma channels.
	- Later when MMC driver resume is called, the driver sends
	  commands to the card. The reception of the response from the
	  card is notified by data completion interrupt.

However, due to a race condition, MMC driver resume callback may be
called even before edma driver resume is called.

Because of this race, MMC driver will be waiting for data completion
interrupt and eventually result in timeout. The response from the card
would have been received but never reached the register in the MMC.

The fix is to move the edma suspend/resume to core PM framework so that
they will be the last in the suspend sequence & 1st in the resume
sequence.

[Hebbar, Gururaja <gururaja.hebbar@ti.com>]
	- Update commit message as per review comments.
	- correct callback function syntax (& remove compile time
	  warning)

Signed-off-by: Arun Joseph <arunjoseph@ti.com>