ARM: OMAP2+: AM33xx: Update CPUidle Code
Current am33xx CPUIdle was wrong in that it was putting EMIF to self
refresh while running code from SDRAM. On AM33xx, EMIF cannot enter Self
refresh in this way.

Also, we observed that cpuidle with EMIF in self refresh causes LCD to
flicker & Audio to cause glitches.

So, instead of putting EMIF to self refresh, another method suggested
was to disable MPU clock when entering CPUidle. This requires A8 to
inform M3 about CPUidle and intern M3 to enable MP clocks when it
receives MPU wake interrupts. So this involves code handling on M3 side.

With this change, Power consumption at runtime is down from ~208mW to
~6.4mW

Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>