i2c: omap: disable adapter class
There is no easy way to disable automatic detection mechanism for
device drivers when a matching class is set for the adapter.

Most embedded devices use a fixed configuration anyway, specifying I²C
addresses in board-...

Automatic detection slows down the boot process and breaks devices if the
detection erroneously "detects" the wrong device, e.g. on piA-AM3517 there
is a LM75 at 0x48 and a 3G sensor (without in kernel driver) at 0x4C. The
3G gets detected as an LM75 and is unusable in this state. Also, this
non-LM75 device introduces other problems when accessed and evaluated as
LM75.

NOTE: This patch should NOT be used for other OMAP devices without
checking that the class based detection mechanism is not used.


Signed-off-by: Björn Krombholz <b.krombholz@pironex.de>